requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.0
supabase>=2.0.0
bcrypt
standard-imghdr